Working Version
Added option to generate relations with no foreign keys. This is done
for performance of read operation reason. Each time row is accessed the
database got hit one time for every foreign key in the table it self and
the tables it refers to.

Added index on all primary keys

Started implementation of adding geometry feature