Split sqlite databases
Now there are three sqlite database files that hold all of the io data.
The Files are defined in db-inventort.txt. Currently those are: Supply,
Demand, Result.

Even though on the back end the IO library worsk with 3 different
databases, it should be seemles for the user of the IO library. The
Sqlite ATTACH mechanism is used to "stich" all of them into a single db
object. Disadvantage is that there is no cross-db foreign keys.

This feature was implemented to split networ and demand so that Network
Editor (http://sourceforge.net/projects/transimsstudio) does not need to
carry around enything by tht esupply related data.